// web/src/app/model/part-dimension/update-part-dimension-model.ts
/* Auto Generated Code By AutoCodeGen Jabil © 2019 */
import { BaseModel } from '../base/base-model';
import { Constants } from 'src/app/shared/constant/global';
import { PartDimension } from './part-dimension';

export class UpdatePartDimensionModel extends BaseModel {
    name: string;
    description: string;

    constructor(partDimension?: PartDimension) {
        super(partDimension);
        if (partDimension) {
            this.name = partDimension.name;
            this.description = partDimension.description;
        } else {
            this.name = Constants.Empty;
            this.description = Constants.Empty;
        }
    }
}